refactor(gaxi): prepare for host reuse
